change the error toast location
