+++ src/pages/Decks/DeckCard.jsx
@@ -1,20 +1,16 @@
-import { Box, Heading, Text, Flex, useColorModeValue, HStack, useToast } from '@chakra-ui/react';
+import { Box, Heading, Text, Flex, useColorModeValue, HStack } from '@chakra-ui/react';
 import { useState } from 'react';
 import { ArrowUpRight, Heart } from 'react-feather';
-import { useNavigate } from 'react-router-dom';
 
 /**
  * This component represents a single deckcard, which consists of the ability to
  * redirect to flashcard page by one click, and a like button on it to save the deckcard
  * @param name
  * @param flashcardCount
- * @param topicId
  * @returns
  */
-export default function DeckCard({ name, flashcardCount, topicId }) {
+export default function DeckCard({ name, flashcardCount, handleError }) {
   const [liked, setLiked] = useState(false);
-  const navigate = useNavigate();
-  const toast = useToast();
 
   return (
     <Box
@@ -51,18 +47,7 @@ export default function DeckCard({ name, flashcardCount, topicId }) {
           roundedBottom={'sm'}
           cursor={'pointer'}
           w='full'
-          onClick={() => {
-            if (flashcardCount) {
-              navigate(`/flashcard/${topicId}`);
-            } else {
-              return toast({
-                title: 'No Flashcards Available',
-                status: 'error',
-                isClosable: true,
-                position: 'top',
-              });
-            }
-          }}
+          onClick={handleError}
         >
           <Text fontSize={'md'} fontWeight={'semibold'}>
             View cards
